requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pandas==2.2.3
numpy==2.2.0
torch==2.0.1
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
python-binance==1.0.19
newsapi-python==0.2.7